Moved the logic of the form from 'new'/'edit' into 'form'
