+++ app/assets/javascript/quotations/quotations.js
@@ -3457,7 +3457,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n\r\n/* harmony default export
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  props:{\n    section_valid: {\n      type: Boolean,\n      default: false\n    },\n    quotation_id:{\n      type:Number,\n      default: null\n    },\n    get_conditions:{\n      type: Boolean,\n      default: false\n    }\n\n  },\n\n  data(){\n    return {\n      translations: I18n.t('quotations.new_edit.conditions'),\n      quotation:{\n        payment_condition:'',\n        credits:'',\n        warranty:''\n      }\n    }\n  },\n\n  mounted(){\n    this.getConditions();\n  },\n\n  methods:{\n    getConditions(){\n      if(this.get_conditions && this.quotation_id){\n        this.http\n        .get(`api/quotations/${this.quotation_id}/conditions`)\n        .then((response)=>{\n          if(response.successful){\n            this.quotation = response.data;\n          }else{\n            this.handleError(response.error);\n          }\n        }).catch((err)=>{\n          console.log(\"Error\", err.stack, err.name, err.message);\n        })\n      }\n    },\n    \n    updateConditions(event){\n      event.preventDefault();\n      this.$emit('update:section_valid', false);\n      this.http\n      .put(`quotations/${this.quotation_id}`, {quotation: this.quotation})\n      .then((response)=>{\n        if(response.successful){\n          this.$emit('update:section_valid', true);\n          this.alert(this.translations.notifications.conditions_updated,'success');\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/conditions.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  props:{\n    section_valid: {\n      type: Boolean,\n      default: false\n    },\n    quotation_id:{\n      type:Number,\n      default: null\n    }\n  },\n\n  data(){\n    return {\n      translations: I18n.t('quotations.new_edit.conditions'),\n      quotation:{\n        payment_condition:'',\n        credits:'',\n        warranty:''\n      }\n    }\n  },\n\n  mounted(){\n    this.getConditions();\n  },\n\n  methods:{\n    getConditions(){\n      if(this.quotation_id){\n        this.http\n        .get(`/api/quotations/${this.quotation_id}/conditions`)\n        .then((response)=>{\n          if(response.successful){\n            this.quotation = response.data;\n          }else{\n            this.handleError(response.error);\n          }\n        }).catch((err)=>{\n          console.log(\"Error\", err.stack, err.name, err.message);\n        })\n      }\n    },\n    \n    updateConditions(event){\n      event.preventDefault();\n      this.$emit('update:section_valid', false);\n      this.http\n      .put(`/quotations/${this.quotation_id}`, {quotation: this.quotation})\n      .then((response)=>{\n        if(response.successful){\n          this.$emit('update:section_valid', true);\n          this.alert(this.translations.notifications.conditions_updated,'success');\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/conditions.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3481,7 +3481,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n\n /* harmony default export 
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  props:{\n    section_valid: {\n      type: Boolean,\n      default: false\n    },\n    quotation_id:{\n      type:Number,\n      default: null\n    },\n    get_header:{\n      type: Boolean,\n      default: false\n    }\n  },\n\n  data(){\n    return {\n      translations: I18n.t('quotations.new_edit.header'),\n      quotation:{\n        client_id: null,\n        client_nit: '',\n        quotation_date: this.date.today(),\n        quotation_type: null\n      },\n      clients: [],\n      quotation_types: []\n    }\n  },\n\n  mounted(){\n    this.getClients();\n    this.getQuotationTypes();\n    this.getHeader();\n  },\n\n  computed:{\n    quotationCode: function(){\n      if(this.quotation_id){\n        return this.quotation_id+100;\n      }\n      return null;\n    }\n  },\n\n  methods:{\n\n    getHeader(){\n      if(this.get_header && this.quotation_id){\n        this.http\n        .get(`api/quotations/${this.quotation_id}/header`)\n        .then((response)=>{\n          if(response.successful){\n            this.quotation = response.data;\n          }else{\n            this.handleError(response.error);\n          }\n        }).catch((err)=>{\n          console.log(\"Error\", err.stack, err.name, err.message);\n        });\n      }\n    },\n\n    getClients(){\n      this.http\n      .get('api/clients')\n      .then((response)=>{\n        if(response.successful){\n          this.clients = response.data;\n          if(!this.get_header && this.clients.length > 0){\n            this.quotation.client_id = this.clients[0].id;\n            this.quotation.client_nit = this.clients[0].nit;\n          }\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    getQuotationTypes(){\n      this.http\n      .get('api/quotations/types')\n      .then((response)=>{\n        if(response.successful){\n          this.quotation_types = response.data;\n          if(!this.get_header && this.quotation_types.length > 0){\n            this.quotation.quotation_type = this.quotation_types[0].value;\n          }\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    submitForm(event){\n      event.preventDefault();\n      if(this.quotation_id){\n        this.updateHeader();\n      }else{\n        this.createHeader();\n      }\n    },\n\n    createHeader(){\n      this.$emit('update:section_valid', false);\n      this.http\n      .post('/quotations', {quotation: this.quotation})\n      .then((response)=>{\n        if(response.successful){\n          this.alert(this.translations.notifications.header_updated,'success');\n          this.$emit('update:quotation_id', response.data.id);\n          this.$emit('update:section_valid', true);\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    updateHeader(){\n      this.$emit('update:section_valid', false);\n      this.http\n      .put(`quotations/${this.quotation_id}`, {quotation: this.quotation})\n      .then((response)=>{\n        if(response.successful){\n          this.alert(this.translations.notifications.header_updated,'success');\n          this.$emit('update:section_valid', true);\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    }\n  },\n\n  watch:{\n    //updates the client's nit when the user changes \n    'quotation.client_id': function(){\n      if(this.clients.length > 0){\n        var selected_client = this.clients.filter((client)=>{\n          return client.id == this.quotation.client_id;\n        });\n        this.quotation.client_nit = selected_client[0].nit;\n      }\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/header.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  props:{\n    section_valid: {\n      type: Boolean,\n      default: false\n    },\n    quotation_id:{\n      type:Number,\n      default: null\n    }\n  },\n\n  data(){\n    return {\n      translations: I18n.t('quotations.new_edit.header'),\n      quotation:{\n        client_id: null,\n        client_nit: '',\n        quotation_date: this.date.today(),\n        quotation_type: null\n      },\n      clients: [],\n      quotation_types: []\n    }\n  },\n\n  mounted(){\n    this.getClients();\n    this.getQuotationTypes();\n    this.getHeader();\n  },\n\n  computed:{\n    quotationCode: function(){\n      if(this.quotation_id){\n        return this.quotation_id+100;\n      }\n      return null;\n    }\n  },\n\n  methods:{\n\n    getHeader(){\n      if(this.get_header){\n        this.http\n        .get(`/api/quotations/${this.quotation_id}/header`)\n        .then((response)=>{\n          if(response.successful){\n            this.quotation = response.data;\n          }else{\n            this.handleError(response.error);\n          }\n        }).catch((err)=>{\n          console.log(\"Error\", err.stack, err.name, err.message);\n        });\n      }\n    },\n\n    getClients(){\n      this.http\n      .get('/api/clients')\n      .then((response)=>{\n        if(response.successful){\n          this.clients = response.data;\n          if(!this.get_header && this.clients.length > 0){\n            this.quotation.client_id = this.clients[0].id;\n            this.quotation.client_nit = this.clients[0].nit;\n          }\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    getQuotationTypes(){\n      this.http\n      .get('/api/quotations/types')\n      .then((response)=>{\n        if(response.successful){\n          this.quotation_types = response.data;\n          if(!this.get_header && this.quotation_types.length > 0){\n            this.quotation.quotation_type = this.quotation_types[0].value;\n          }\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    submitForm(event){\n      event.preventDefault();\n      if(this.quotation_id){\n        this.updateHeader();\n      }else{\n        this.createHeader();\n      }\n    },\n\n    createHeader(){\n      this.$emit('update:section_valid', false);\n      this.http\n      .post('/quotations', {quotation: this.quotation})\n      .then((response)=>{\n        if(response.successful){\n          this.alert(this.translations.notifications.header_updated,'success');\n          this.$emit('update:quotation_id', response.data.id);\n          this.$emit('update:section_valid', true);\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    updateHeader(){\n      this.$emit('update:section_valid', false);\n      this.http\n      .put(`/quotations/${this.quotation_id}`, {quotation: this.quotation})\n      .then((response)=>{\n        if(response.successful){\n          this.alert(this.translations.notifications.header_updated,'success');\n          this.$emit('update:section_valid', true);\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    }\n  },\n\n  watch:{\n    //updates the client's nit when the user changes \n    'quotation.client_id': function(){\n      if(this.clients.length > 0){\n        var selected_client = this.clients.filter((client)=>{\n          return client.id == this.quotation.client_id;\n        });\n        this.quotation.client_nit = selected_client[0].nit;\n      }\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/header.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3493,7 +3493,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n\r\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\r\n  props: {\r\n    section_valid: {\r\n      type: Boolean,\r\n      default: false\r\n    },\r\n    quotation_id: {\r\n      type: Number,\r\n      default: null\r\n    },\r\n    get_products:{\r\n      type: Boolean,\r\n      default: false\r\n    }\r\n  },\r\n  data() {\r\n    return {\r\n      translations: I18n.t('quotations.new_edit.materials'),\r\n      quotation_products: {\r\n        amount: 1,\r\n        percents: [15.0,15.0]\r\n      },\r\n      quotation_type: null,\r\n      material_id: null,\r\n      materials: [],\r\n      products: [],\r\n      prices: [],\r\n      brands: [],\r\n      selected_materials: [],\r\n      deleted_materials: [],\r\n      table_headers: [],\r\n      product_id: null\r\n    };\r\n  },\r\n  mounted(){\r\n    this.getQuotationType();\r\n    this.getQuotationProducts();\r\n  },\r\n  methods: {\r\n    getQuotationProducts: function(){\r\n      if(this.get_products && this.quotation_id){\r\n        this.http\r\n        .get(`api/quotations/${this.quotation_id}/products`)\r\n        .then((response)=>{\r\n          if(response.successful){\r\n            this.selected_materials = response.data;\r\n          }else{\r\n            this.handleError(response.error);\r\n          }\r\n        }).catch((err)=>{\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n      }\r\n    },\r\n\r\n    setTableHeaders: function(){\r\n      switch(this.quotation_type){\r\n        case 't_comparative': \r\n        ['amount',\r\n          'material',\r\n          'percent_supranet',\r\n          'price_supranet',\r\n          'total_supranet',\r\n          'price_percent_supranet',\r\n          'total_percent_supranet',\r\n          'percent_siemon',\r\n          'price_siemon',\r\n          'total_siemon',\r\n          'price_percent_siemon',\r\n          'total_percent_siemon'].forEach((element)=>{\r\n            this.table_headers.push({\r\n              key: element,\r\n              label: this.translations.headers[element]\r\n            });\r\n          });\r\n        break;\r\n        case 't_simple':\r\n          ['amount','code','material','brand','percent','price','total','price_percent','total_percent'].forEach((element)=>{\r\n            this.table_headers.push({\r\n              key:element,\r\n              label:this.translations.headers[element]\r\n            });\r\n          });   \r\n        break;\r\n        default: \r\n          ['amount','code','material','percent','price','total','price_percent','total_percent'].forEach((element)=>{\r\n            this.table_headers.push({\r\n              key:element,\r\n              label:this.translations.headers[element]\r\n            });\r\n          });    \r\n        break;\r\n      }\r\n      this.table_headers.push({actions: ''});\r\n    },\r\n    comparativeMaterials: function(){\r\n      this.http\r\n      .get(\"/api/comparative\")\r\n      .then(response => {\r\n        if(response.successful){\r\n          this.materials = response.data;\r\n          this.materials = this.materials.map(function(material) {\r\n            return {\r\n              id: material.id,\r\n              name: `${material.name} ${material.description}`\r\n            };\r\n          });\r\n          if (this.materials.length > 0) {\r\n            this.material_id = this.materials[0].id;\r\n          }\r\n        }else{\r\n          this.handleError(response.error);\r\n        }\r\n      })\r\n      .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n      });\r\n    },\r\n    simpleMaterials: function(){\r\n      this.http\r\n      .get(\"/api/simple\")\r\n      .then(response =>{\r\n        if(response.successful){\r\n          this.materials = response.data;\r\n          this.materials = this.materials.map(function(material){\r\n            return{\r\n              id: material.product_id,\r\n              name: `${material.brand} - ${material.name} ${material.description}`,\r\n              material_id: material.material_id,\r\n              product_id: material.product_id,\r\n              brand: material.brand,\r\n              code: material.code,\r\n              price: material.price,\r\n              measure_unit: material.measure_unit\r\n            }\r\n          });\r\n          if (this.materials.length > 0) {\r\n            this.material_id = this.materials[0].id;\r\n            this.prices[0] = this.materials[0].price;\r\n          }\r\n        }else{\r\n          this.handleError(response.error);\r\n        }\r\n      })\r\n      .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n      });\r\n    },\r\n    onlyBrandMaterials: function(){\r\n      this.http\r\n      .get(\"/api/simple\")\r\n      .then(response =>{\r\n        if(response.successful){\r\n          this.materials = response.data;\r\n          this.materials = this.materials.filter(material => material.brand !== this.brands[1]);\r\n          this.materials = this.materials.map(function(material){\r\n            return{\r\n              id: material.material_id,\r\n              name: `${material.name} ${material.description}`,\r\n              product_id: material.product_id,\r\n              brand: material.brand,\r\n              code: material.code,\r\n              price: material.price,\r\n              measure_unit: material.measure_unit\r\n            }\r\n          });\r\n          if (this.materials.length > 0) {\r\n            this.material_id = this.materials[0].id;\r\n          }\r\n        }else{\r\n          this.handleError(response.error);\r\n        }\r\n      })\r\n      .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n      });\r\n    },\r\n    getMaterials: function() {\r\n      if (this.quotation_type === 't_comparative'){\r\n        this.comparativeMaterials();\r\n      }else if(this.quotation_type === 't_simple'){\r\n        this.simpleMaterials();\r\n      }else{\r\n        this.onlyBrandMaterials();\r\n      }\r\n    },\r\n    getProducts: function() {\r\n      this.http\r\n        .get(\"/api/products_by_material\", {\r\n          params: {\r\n            material_id: this.material_id\r\n          }\r\n        })\r\n        .then(response => {\r\n          if(response.successful){\r\n            this.products = response.data;\r\n            if(this.quotation_type!=='t_comparative'){\r\n              this.products = this.products.filter(product => product.brand === this.brands[0]);\r\n            }\r\n            this.setPrices();\r\n          }else{\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n    },\r\n    setPrices: function() {\r\n      this.prices = this.products.map(function(product) {\r\n        return product.price;\r\n      });\r\n      if (this.quotation_type === \"t_comparative\" && this.prices.length === 1) {\r\n        this.prices[1] = this.prices[0];\r\n      }\r\n    },\r\n    getQuotationType: function() {\r\n      if(this.quotation_id){\r\n        this.http\r\n          .get(`/api/quotations/${this.quotation_id}/type`)\r\n          .then(response => {\r\n            if(response.successful){\r\n              this.quotation_type = response.data[0].quotation_type;\r\n              switch (this.quotation_type) {\r\n                case \"t_comparative\":\r\n                  this.brands = [\"Supranet\", \"Siemon\"];\r\n                  break;\r\n                case \"t_siemon_only\":\r\n                  this.brands = [\"Siemon\",\"Supranet\"];\r\n                  break;\r\n                case \"t_supranet_only\":\r\n                  this.brands = [\"Supranet\",\"Siemon\"];\r\n                  break;\r\n                case \"t_simple\":\r\n                  this.brands = [null];\r\n                  break;\r\n              }\r\n              this.getMaterials();\r\n            }else{\r\n              this.handleError(response.error);\r\n            }\r\n          })\r\n          .catch(err => {\r\n            console.log(\"Error\", err.stack, err.name, err.message);\r\n          });\r\n      }\r\n    },\r\n    addComparativeProduct: function(){\r\n      var price_supranet = this.prices[0];\r\n      var price_siemon = this.prices[1];\r\n      var total_supranet = this.quotation_products.amount * price_supranet;\r\n      var total_siemon = this.quotation_products.amount * price_siemon;\r\n      var total_percent_supranet = (total_supranet * this.percentage.format(this.quotation_products.percents[0]));\r\n      var total_percent_siemon = (total_siemon * this.percentage.format(this.quotation_products.percents[1]));\r\n      // Obtaind id for Second product\r\n      var siemon_id = null;\r\n      if (this.products.length>1){\r\n        siemon_id = this.products[1].product_id;\r\n      }\r\n      // Material name\r\n      var material = this.materials.filter(material => material.id === this.material_id);\r\n\r\n      this.selected_materials.push({\r\n        id: this.product_id,\r\n        material_id: this.material_id,\r\n        amount: `${this.quotation_products.amount}`,\r\n        material: `${material[0].name}`,\r\n        supranet_id: `${this.products[0].product_id}`,\r\n        percent_supranet: `${this.quotation_products.percents[0]}`,\r\n        price_supranet: price_supranet,\r\n        total_supranet: `${total_supranet.toFixed(2)}`,\r\n        price_percent_supranet: price_supranet,\r\n        total_percent_supranet: total_percent_supranet.toFixed(2),\r\n        siemon_id: siemon_id,\r\n        percent_siemon: `${this.quotation_products.percents[1]}`,\r\n        price_siemon: price_siemon,\r\n        total_siemon: `${total_siemon.toFixed(2)}`,\r\n        price_percent_siemon: price_siemon,\r\n        total_percent_siemon: total_percent_siemon.toFixed(2)\r\n      });\r\n    },\r\n    addSimpleProduct: function(){\r\n      var material = this.materials.filter(material => material.id === this.material_id);\r\n      var price = material[0].price;\r\n      var total = price * this.quotation_products.amount;\r\n      var total_percent = total * this.percentage.format(this.quotation_products.percents[0]);\r\n      this.selected_materials.push({\r\n        id: this.product_id,\r\n        material_id: this.material_id,\r\n        amount: this.quotation_products.amount,\r\n        code: `${material[0].code}`,\r\n        material: `${material[0].name}`,\r\n        brand: `${material[0].brand}`,\r\n        product_id: material[0].product_id,\r\n        percent: this.quotation_products.percents[0],\r\n        price: price,\r\n        total: this.currency.format(total),\r\n        price_percent: price,\r\n        total_percent: this.currency.format(total_percent),\r\n      });\r\n    },\r\n    addProducts: function() {\r\n      var add_product = this.selected_materials.filter(selected => selected.material_id === this.material_id)\r\n      if (add_product.length===0){\r\n        switch(this.quotation_type){\r\n          case 't_comparative':\r\n            this.addComparativeProduct();\r\n            break;\r\n          case 't_simple':\r\n            this.addSimpleProduct();\r\n            break;\r\n          default: \r\n            this.addSimpleProduct();\r\n            break;\r\n        }\r\n        this.product_id = null;\r\n      }\r\n    },\r\n    editProduct: function(index) {\r\n      let is_comparative = this.quotation_type === 't_comparative';\r\n      let product_data = this.deleteProduct(index,is_comparative);\r\n      if(is_comparative){\r\n        this.quotation_products.percents[0] = product_data.percent_supranet;\r\n        this.quotation_products.percents[1] = product_data.percent_siemon;\r\n        this.prices[0] = product_data.price_supranet;\r\n        this.prices[1] = product_data.price_siemon;\r\n      }else{\r\n        this.quotation_products.percents[0] = product_data.percent;\r\n        this.prices[0] = product_data.price;\r\n      }\r\n      this.quotation_products.amount = product_data.amount;\r\n      this.material_id = product_data.material_id;\r\n      //The ID of the product in the database, if there is any\r\n      this.product_id = product_data.id;\r\n    },\r\n    deleteProduct: function(index,staged_for_delete=true) {\r\n      //If there is and ID, we have to delete the record at the database\r\n      var quotation_product = this.selected_materials.splice(index, 1)[0];\r\n      if(staged_for_delete){\r\n        ['quotation_product_supranet_id','quotation_product_siemon_id','id'].forEach((id)=>{\r\n          if(quotation_product[id]){\r\n            this.deleted_materials.push({id: quotation_product[id], _destroy: true});\r\n          }\r\n        });\r\n      }\r\n      return quotation_product;\r\n    },\r\n    formatData: function() {\r\n      var data;\r\n      if (this.quotation_type === \"t_comparative\") {\r\n        var product_attributes = [];\r\n        this.selected_materials.forEach(material => {\r\n          product_attributes.push({\r\n            id: material.quotation_product_supranet_id,\r\n            amount: material.amount,\r\n            percent: material.percent_supranet,\r\n            product_id: material.supranet_id\r\n          });\r\n          if (material.siemon_id) {\r\n            product_attributes.push({\r\n              id: material.quotation_product_siemon_id,\r\n              amount: material.amount,\r\n              percent: material.percent_siemon,\r\n              product_id: material.siemon_id\r\n            });\r\n          }\r\n        });\r\n        data = {\r\n          quotation: {\r\n            quotation_products_attributes: product_attributes.concat(this.deleted_materials)\r\n          }\r\n        };\r\n      }else{\r\n        data = {\r\n          quotation: {\r\n            quotation_products_attributes: this.selected_materials.concat(this.deleted_materials)\r\n          }\r\n        };\r\n      }\r\n      return data;\r\n    },\r\n    submit: function() {\r\n      this.$emit(\"update:section_valid\", false);\r\n      this.http\r\n        .put(`api/quotations/${this.quotation_id}/update`, this.formatData())\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.$emit(\"update:section_valid\", true);\r\n            this.alert(this.translations.notifications.materials_updated,'success');\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n    }\r\n  },\r\n  watch: {\r\n    material_id: function() {\r\n      if(this.quotation_type === 't_comparative'){\r\n        this.getProducts();\r\n        this.percents =[15.0,15.0];\r\n        this.amount = 1;\r\n      }else{\r\n        var material = this.materials.filter(material => material.id === this.material_id);\r\n        this.prices[0] = material[0].price;\r\n        this.percents =[15.0,15.0];\r\n        this.amount = 1;\r\n      }\r\n    },\r\n    quotation_id: function() {\r\n      this.getQuotationType();\r\n    },\r\n    quotation_type: function(){\r\n      this.setTableHeaders();\r\n    }\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/products.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n\r\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\r\n  props: {\r\n    section_valid: {\r\n      type: Boolean,\r\n      default: false\r\n    },\r\n    quotation_id: {\r\n      type: Number,\r\n      default: null\r\n    }\r\n  },\r\n  data() {\r\n    return {\r\n      translations: I18n.t('quotations.new_edit.materials'),\r\n      quotation_products: {\r\n        amount: 1,\r\n        percents: [15.0,15.0]\r\n      },\r\n      quotation_type: null,\r\n      material_id: null,\r\n      materials: [],\r\n      products: [],\r\n      prices: [],\r\n      brands: [],\r\n      selected_materials: [],\r\n      deleted_materials: [],\r\n      table_headers: [],\r\n      product_id: null\r\n    };\r\n  },\r\n  mounted(){\r\n    this.getQuotationType();\r\n    this.getQuotationProducts();\r\n  },\r\n  methods: {\r\n    getQuotationProducts: function(){\r\n      if(this.quotation_id){\r\n        this.http\r\n        .get(`/api/quotations/${this.quotation_id}/products`)\r\n        .then((response)=>{\r\n          if(response.successful){\r\n            this.selected_materials = response.data;\r\n          }else{\r\n            this.handleError(response.error);\r\n          }\r\n        }).catch((err)=>{\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n      }\r\n    },\r\n\r\n    setTableHeaders: function(){\r\n      switch(this.quotation_type){\r\n        case 't_comparative': \r\n        ['amount',\r\n          'material',\r\n          'percent_supranet',\r\n          'price_supranet',\r\n          'total_supranet',\r\n          'price_percent_supranet',\r\n          'total_percent_supranet',\r\n          'percent_siemon',\r\n          'price_siemon',\r\n          'total_siemon',\r\n          'price_percent_siemon',\r\n          'total_percent_siemon'].forEach((element)=>{\r\n            this.table_headers.push({\r\n              key: element,\r\n              label: this.translations.headers[element]\r\n            });\r\n          });\r\n        break;\r\n        case 't_simple':\r\n          ['amount','code','material','brand','percent','price','total','price_percent','total_percent'].forEach((element)=>{\r\n            this.table_headers.push({\r\n              key:element,\r\n              label:this.translations.headers[element]\r\n            });\r\n          });   \r\n        break;\r\n        default: \r\n          ['amount','code','material','percent','price','total','price_percent','total_percent'].forEach((element)=>{\r\n            this.table_headers.push({\r\n              key:element,\r\n              label:this.translations.headers[element]\r\n            });\r\n          });    \r\n        break;\r\n      }\r\n      this.table_headers.push({actions: ''});\r\n    },\r\n    comparativeMaterials: function(){\r\n      this.http\r\n      .get(\"/api/comparative\")\r\n      .then(response => {\r\n        if(response.successful){\r\n          this.materials = response.data;\r\n          this.materials = this.materials.map(function(material) {\r\n            return {\r\n              id: material.id,\r\n              name: `${material.name} ${material.description}`\r\n            };\r\n          });\r\n          if (this.materials.length > 0) {\r\n            this.material_id = this.materials[0].id;\r\n          }\r\n        }else{\r\n          this.handleError(response.error);\r\n        }\r\n      })\r\n      .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n      });\r\n    },\r\n    simpleMaterials: function(){\r\n      this.http\r\n      .get(\"/api/simple\")\r\n      .then(response =>{\r\n        if(response.successful){\r\n          this.materials = response.data;\r\n          this.materials = this.materials.map(function(material){\r\n            return{\r\n              id: material.product_id,\r\n              name: `${material.brand} - ${material.name} ${material.description}`,\r\n              material_id: material.material_id,\r\n              product_id: material.product_id,\r\n              brand: material.brand,\r\n              code: material.code,\r\n              price: material.price,\r\n              measure_unit: material.measure_unit\r\n            }\r\n          });\r\n          if (this.materials.length > 0) {\r\n            this.material_id = this.materials[0].id;\r\n            this.prices[0] = this.materials[0].price;\r\n          }\r\n        }else{\r\n          this.handleError(response.error);\r\n        }\r\n      })\r\n      .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n      });\r\n    },\r\n    onlyBrandMaterials: function(){\r\n      this.http\r\n      .get(\"/api/simple\")\r\n      .then(response =>{\r\n        if(response.successful){\r\n          this.materials = response.data;\r\n          this.materials = this.materials.filter(material => material.brand !== this.brands[1]);\r\n          this.materials = this.materials.map(function(material){\r\n            return{\r\n              id: material.material_id,\r\n              name: `${material.name} ${material.description}`,\r\n              product_id: material.product_id,\r\n              brand: material.brand,\r\n              code: material.code,\r\n              price: material.price,\r\n              measure_unit: material.measure_unit\r\n            }\r\n          });\r\n          if (this.materials.length > 0) {\r\n            this.material_id = this.materials[0].id;\r\n          }\r\n        }else{\r\n          this.handleError(response.error);\r\n        }\r\n      })\r\n      .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n      });\r\n    },\r\n    getMaterials: function() {\r\n      if (this.quotation_type === 't_comparative'){\r\n        this.comparativeMaterials();\r\n      }else if(this.quotation_type === 't_simple'){\r\n        this.simpleMaterials();\r\n      }else{\r\n        this.onlyBrandMaterials();\r\n      }\r\n    },\r\n    getProducts: function() {\r\n      this.http\r\n        .get(\"/api/products_by_material\", {\r\n          params: {\r\n            material_id: this.material_id\r\n          }\r\n        })\r\n        .then(response => {\r\n          if(response.successful){\r\n            this.products = response.data;\r\n            if(this.quotation_type!=='t_comparative'){\r\n              this.products = this.products.filter(product => product.brand === this.brands[0]);\r\n            }\r\n            this.setPrices();\r\n          }else{\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n        console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n    },\r\n    setPrices: function() {\r\n      this.prices = this.products.map(function(product) {\r\n        return product.price;\r\n      });\r\n      if (this.quotation_type === \"t_comparative\" && this.prices.length === 1) {\r\n        this.prices[1] = this.prices[0];\r\n      }\r\n    },\r\n    getQuotationType: function() {\r\n      if(this.quotation_id){\r\n        this.http\r\n          .get(`/api/quotations/${this.quotation_id}/type`)\r\n          .then(response => {\r\n            if(response.successful){\r\n              this.quotation_type = response.data[0].quotation_type;\r\n              switch (this.quotation_type) {\r\n                case \"t_comparative\":\r\n                  this.brands = [\"Supranet\", \"Siemon\"];\r\n                  break;\r\n                case \"t_siemon_only\":\r\n                  this.brands = [\"Siemon\",\"Supranet\"];\r\n                  break;\r\n                case \"t_supranet_only\":\r\n                  this.brands = [\"Supranet\",\"Siemon\"];\r\n                  break;\r\n                case \"t_simple\":\r\n                  this.brands = [null];\r\n                  break;\r\n              }\r\n              this.getMaterials();\r\n            }else{\r\n              this.handleError(response.error);\r\n            }\r\n          })\r\n          .catch(err => {\r\n            console.log(\"Error\", err.stack, err.name, err.message);\r\n          });\r\n      }\r\n    },\r\n    addComparativeProduct: function(){\r\n      var price_supranet = this.prices[0];\r\n      var price_siemon = this.prices[1];\r\n      var total_supranet = this.quotation_products.amount * price_supranet;\r\n      var total_siemon = this.quotation_products.amount * price_siemon;\r\n      var total_percent_supranet = (total_supranet * this.percentage.format(this.quotation_products.percents[0]));\r\n      var total_percent_siemon = (total_siemon * this.percentage.format(this.quotation_products.percents[1]));\r\n      // Obtaind id for Second product\r\n      var siemon_id = null;\r\n      if (this.products.length>1){\r\n        siemon_id = this.products[1].product_id;\r\n      }\r\n      // Material name\r\n      var material = this.materials.filter(material => material.id === this.material_id);\r\n\r\n      this.selected_materials.push({\r\n        id: this.product_id,\r\n        material_id: this.material_id,\r\n        amount: `${this.quotation_products.amount}`,\r\n        material: `${material[0].name}`,\r\n        supranet_id: `${this.products[0].product_id}`,\r\n        percent_supranet: `${this.quotation_products.percents[0]}`,\r\n        price_supranet: price_supranet,\r\n        total_supranet: `${total_supranet.toFixed(2)}`,\r\n        price_percent_supranet: price_supranet,\r\n        total_percent_supranet: total_percent_supranet.toFixed(2),\r\n        siemon_id: siemon_id,\r\n        percent_siemon: `${this.quotation_products.percents[1]}`,\r\n        price_siemon: price_siemon,\r\n        total_siemon: `${total_siemon.toFixed(2)}`,\r\n        price_percent_siemon: price_siemon,\r\n        total_percent_siemon: total_percent_siemon.toFixed(2)\r\n      });\r\n    },\r\n    addSimpleProduct: function(){\r\n      var material = this.materials.filter(material => material.id === this.material_id);\r\n      var price = material[0].price;\r\n      var total = price * this.quotation_products.amount;\r\n      var total_percent = total * this.percentage.format(this.quotation_products.percents[0]);\r\n      this.selected_materials.push({\r\n        id: this.product_id,\r\n        material_id: this.material_id,\r\n        amount: this.quotation_products.amount,\r\n        code: `${material[0].code}`,\r\n        material: `${material[0].name}`,\r\n        brand: `${material[0].brand}`,\r\n        product_id: material[0].product_id,\r\n        percent: this.quotation_products.percents[0],\r\n        price: price,\r\n        total: this.currency.format(total),\r\n        price_percent: price,\r\n        total_percent: this.currency.format(total_percent),\r\n      });\r\n    },\r\n    addProducts: function() {\r\n      var add_product = this.selected_materials.filter(selected => selected.material_id === this.material_id)\r\n      if (add_product.length===0){\r\n        switch(this.quotation_type){\r\n          case 't_comparative':\r\n            this.addComparativeProduct();\r\n            break;\r\n          case 't_simple':\r\n            this.addSimpleProduct();\r\n            break;\r\n          default: \r\n            this.addSimpleProduct();\r\n            break;\r\n        }\r\n        this.product_id = null;\r\n      }\r\n    },\r\n    editProduct: function(index) {\r\n      let is_comparative = this.quotation_type === 't_comparative';\r\n      let product_data = this.deleteProduct(index,is_comparative);\r\n      if(is_comparative){\r\n        this.quotation_products.percents[0] = product_data.percent_supranet;\r\n        this.quotation_products.percents[1] = product_data.percent_siemon;\r\n        this.prices[0] = product_data.price_supranet;\r\n        this.prices[1] = product_data.price_siemon;\r\n      }else{\r\n        this.quotation_products.percents[0] = product_data.percent;\r\n        this.prices[0] = product_data.price;\r\n      }\r\n      this.quotation_products.amount = product_data.amount;\r\n      this.material_id = product_data.material_id;\r\n      //The ID of the product in the database, if there is any\r\n      this.product_id = product_data.id;\r\n    },\r\n    deleteProduct: function(index,staged_for_delete=true) {\r\n      //If there is and ID, we have to delete the record at the database\r\n      var quotation_product = this.selected_materials.splice(index, 1)[0];\r\n      if(staged_for_delete){\r\n        ['quotation_product_supranet_id','quotation_product_siemon_id','id'].forEach((id)=>{\r\n          if(quotation_product[id]){\r\n            this.deleted_materials.push({id: quotation_product[id], _destroy: true});\r\n          }\r\n        });\r\n      }\r\n      return quotation_product;\r\n    },\r\n    formatData: function() {\r\n      var data;\r\n      if (this.quotation_type === \"t_comparative\") {\r\n        var product_attributes = [];\r\n        this.selected_materials.forEach(material => {\r\n          product_attributes.push({\r\n            id: material.quotation_product_supranet_id,\r\n            amount: material.amount,\r\n            percent: material.percent_supranet,\r\n            product_id: material.supranet_id\r\n          });\r\n          if (material.siemon_id) {\r\n            product_attributes.push({\r\n              id: material.quotation_product_siemon_id,\r\n              amount: material.amount,\r\n              percent: material.percent_siemon,\r\n              product_id: material.siemon_id\r\n            });\r\n          }\r\n        });\r\n        data = {\r\n          quotation: {\r\n            quotation_products_attributes: product_attributes.concat(this.deleted_materials)\r\n          }\r\n        };\r\n      }else{\r\n        data = {\r\n          quotation: {\r\n            quotation_products_attributes: this.selected_materials.concat(this.deleted_materials)\r\n          }\r\n        };\r\n      }\r\n      return data;\r\n    },\r\n    submit: function() {\r\n      this.$emit(\"update:section_valid\", false);\r\n      this.http\r\n        .put(`/api/quotations/${this.quotation_id}/update`, this.formatData())\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.$emit(\"update:section_valid\", true);\r\n            this.alert(this.translations.notifications.materials_updated,'success');\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n    }\r\n  },\r\n  watch: {\r\n    material_id: function() {\r\n      if(this.quotation_type === 't_comparative'){\r\n        this.getProducts();\r\n        this.percents =[15.0,15.0];\r\n        this.amount = 1;\r\n      }else{\r\n        var material = this.materials.filter(material => material.id === this.material_id);\r\n        this.prices[0] = material[0].price;\r\n        this.percents =[15.0,15.0];\r\n        this.amount = 1;\r\n      }\r\n    },\r\n    quotation_id: function() {\r\n      this.getQuotationType();\r\n    },\r\n    quotation_type: function(){\r\n      this.setTableHeaders();\r\n    }\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/products.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3505,7 +3505,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n\r\n/* harmony default export
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  props:{\n    section_valid: {\n      type: Boolean,\n      default: false\n    },\n    quotation_id:{\n      type:Number,\n      default: null\n    },\n    get_services:{\n      type: Boolean,\n      default: false\n    }\n  },\n  \n  data() {\n    return {\n      translations: I18n.t('quotations.new_edit.services'),\n      form_fields: {\n        service_id: 0,\n        price: 0,\n        amount: 1,\n        percent: 0\n      },\n      services: [],\n      quotation_services: [],\n      deleted_quotation_services: [],\n      table_headers: []\n    }\n  },\n\n  mounted() {\n    this.getService();\n    this.setTableHeaders();\n    this.getQuotationServices();\n  },\n  \n  methods: {\n    //When the quotation is in 'created' state, this method gets the services that have\n    //been already saved and shows them\n    getQuotationServices(){\n      if(this.get_services && this.quotation_id){\n        this.http\n        .get(`api/quotations/${this.quotation_id}/services`)\n        .then((response)=>{\n          if(response.successful){\n            this.quotation_services = response.data;\n          }else{\n            this.handleError(response.error);\n          }\n        }).catch((err)=>{\n          console.log(\"Error\", err.stack, err.name, err.message);\n        });\n      }\n    },\n\n    getService() {\n      this.http\n      .get('/api/services')\n      .then((response)=>{\n        if(response.successful){\n          this.services = response.data;\n          if(this.services.length > 0){\n            this.form_fields.service_id = this.services[0].id;\n            this.form_fields.price = this.services[0].price;\n          }\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    setTableHeaders() {\n      ['name','price','amount','percent','tot_without_perc','tot_with_perc','actions'].forEach((element)=>{\n        this.table_headers.push({\n          key:element,\n          label:this.translations.headers[element],\n          sortable:true\n        })\n      });\n    },\n\n    addService(event) {\n      event.preventDefault();\n      var form_data = this.form_fields;\n      var selected_service = this.services.filter((service)=>{\n        return service.id == form_data.service_id;\n      });\n      var tot_without_perc = form_data.amount * form_data.price;\n      //Using ES6 to merge form_data and some extra fields into table_data\n      var table_data = { ...form_data, ...{\n        tot_without_perc: tot_without_perc,\n        tot_with_perc: tot_without_perc * (1+form_data.percent/100),\n        name: selected_service[0].name\n      }};\n      //Remove the ID field if it exists\n      delete this.form_fields.id;\n      this.quotation_services.push(table_data);\n    },\n\n    updateServices(event) {\n      event.preventDefault();\n      this.$emit('update:section_valid', false);\n      var data = {\n        quotation:{\n          quotation_services_attributes: this.quotation_services.concat(this.deleted_quotation_services)\n        }\n      };\n      this.deleted_quotation_services = [];\n      this.http\n      .put(`quotations/${this.quotation_id}`, data)\n      .then((response)=>{\n        if(response.successful){\n          this.alert(this.translations.notifications.services_updated,'success');\n          this.$emit('update:section_valid', true);\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n    \n    deleteService: function(table_row_data) {\n      var quotation_service = this.quotation_services.splice(table_row_data.index, 1)[0];\n      //If there is and ID, we have to delete it at the database\n      if(quotation_service.id){\n        this.deleted_quotation_services.push({id: quotation_service.id, _destroy: true})\n      }\n    },\n\n    editService: function(table_row_data) {\n      this.form_fields = table_row_data.item;\n      this.quotation_services.splice(table_row_data.index, 1);\n    },\n  },\n  \n  watch: {\n    'form_fields.service_id': function(){\n      var selected_service = this.services.filter((service)=>{\n        return service.id == this.form_fields.service_id;\n      });\n      this.form_fields.price  = selected_service[0].price;\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/services.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  props:{\n    section_valid: {\n      type: Boolean,\n      default: false\n    },\n    quotation_id:{\n      type:Number,\n      default: null\n    }\n  },\n  \n  data() {\n    return {\n      translations: I18n.t('quotations.new_edit.services'),\n      form_fields: {\n        service_id: 0,\n        price: 0,\n        amount: 1,\n        percent: 0\n      },\n      services: [],\n      quotation_services: [],\n      deleted_quotation_services: [],\n      table_headers: []\n    }\n  },\n\n  mounted() {\n    this.getService();\n    this.setTableHeaders();\n    this.getQuotationServices();\n  },\n  \n  methods: {\n    //When the quotation is in 'created' state, this method gets the services that have\n    //been already saved and shows them\n    getQuotationServices(){\n      if(this.quotation_id){\n        this.http\n        .get(`api/quotations/${this.quotation_id}/services`)\n        .then((response)=>{\n          if(response.successful){\n            this.quotation_services = response.data;\n          }else{\n            this.handleError(response.error);\n          }\n        }).catch((err)=>{\n          console.log(\"Error\", err.stack, err.name, err.message);\n        });\n      }\n    },\n\n    getService() {\n      this.http\n      .get('/api/services')\n      .then((response)=>{\n        if(response.successful){\n          this.services = response.data;\n          if(this.services.length > 0){\n            this.form_fields.service_id = this.services[0].id;\n            this.form_fields.price = this.services[0].price;\n          }\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n\n    setTableHeaders() {\n      ['name','price','amount','percent','tot_without_perc','tot_with_perc','actions'].forEach((element)=>{\n        this.table_headers.push({\n          key:element,\n          label:this.translations.headers[element],\n          sortable:true\n        })\n      });\n    },\n\n    addService(event) {\n      event.preventDefault();\n      var form_data = this.form_fields;\n      var selected_service = this.services.filter((service)=>{\n        return service.id == form_data.service_id;\n      });\n      var tot_without_perc = form_data.amount * form_data.price;\n      //Using ES6 to merge form_data and some extra fields into table_data\n      var table_data = { ...form_data, ...{\n        tot_without_perc: tot_without_perc,\n        tot_with_perc: tot_without_perc * (1+form_data.percent/100),\n        name: selected_service[0].name\n      }};\n      //Remove the ID field if it exists\n      delete this.form_fields.id;\n      this.quotation_services.push(table_data);\n    },\n\n    updateServices(event) {\n      event.preventDefault();\n      this.$emit('update:section_valid', false);\n      var data = {\n        quotation:{\n          quotation_services_attributes: this.quotation_services.concat(this.deleted_quotation_services)\n        }\n      };\n      this.deleted_quotation_services = [];\n      this.http\n      .put(`quotations/${this.quotation_id}`, data)\n      .then((response)=>{\n        if(response.successful){\n          this.alert(this.translations.notifications.services_updated,'success');\n          this.$emit('update:section_valid', true);\n        }else{\n          this.handleError(response.error);\n        }\n      }).catch((err)=>{\n        console.log(\"Error\", err.stack, err.name, err.message);\n      });\n    },\n    \n    deleteService: function(table_row_data) {\n      var quotation_service = this.quotation_services.splice(table_row_data.index, 1)[0];\n      //If there is and ID, we have to delete it at the database\n      if(quotation_service.id){\n        this.deleted_quotation_services.push({id: quotation_service.id, _destroy: true})\n      }\n    },\n\n    editService: function(table_row_data) {\n      this.form_fields = table_row_data.item;\n      this.quotation_services.splice(table_row_data.index, 1);\n    },\n  },\n  \n  watch: {\n    'form_fields.service_id': function(){\n      var selected_service = this.services.filter((service)=>{\n        return service.id == this.form_fields.service_id;\n      });\n      this.form_fields.price  = selected_service[0].price;\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/services.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3517,7 +3517,19 @@ eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _components_header_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./components/header.vue */ \"./viewsrc/modules/quotations/components/header.vue\");\n/* harmony import */ var _components_products_vue__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/products.vue */ \"./viewsrc/modules/quotations/components/products.vue\");\n/* harmony import */ var _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/conditions.vue */ \"./viewsrc/modules/quotations/components/conditions.vue\");\n/* harmony import */ var _components_services_vue__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/services.vue */ \"./viewsrc/modules/quotations/components/services.vue\");\n/* harmony import */ var _components_global_view_vue__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./components/global_view.vue */ \"./viewsrc/modules/quotations/components/global_view.vue\");\n/* harmony import */ var _components_attachments_vue__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./components/attachments.vue */ \"./viewsrc/modules/quotations/components/attachments.vue\");\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\r\n  components: {\r\n    \"quotation-header\": _components_header_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"],\r\n    \"quotation-products\": _components_products_vue__WEBPACK_IMPORTED_MODULE_1__[\"default\"],\r\n    \"quotation-services\": _components_services_vue__WEBPACK_IMPORTED_MODULE_3__[\"default\"],\r\n    \"quotation-conditions\": _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__[\"default\"],\r\n    \"quotation-global-view\": _components_global_view_vue__WEBPACK_IMPORTED_MODULE_4__[\"default\"],\r\n    \"quotation-attachments\": _components_attachments_vue__WEBPACK_IMPORTED_MODULE_5__[\"default\"]\r\n  },\r\n\r\n  data() {\r\n    return {\r\n      translations: I18n.t(\"quotations.new_edit\"),\r\n      global_view_index: 4,\r\n      tab_index: 4,\r\n      //these are only used to change the color of the tabs to green when the user finishes a section\r\n      completed: {\r\n        header: true,\r\n        products: true,\r\n        services: true,\r\n        conditions: true,\r\n        global_view: false\r\n      },\r\n      refresh_global_view: false,\r\n      quotation: {\r\n        id: parseInt(this.$route.params.id)\r\n      }\r\n    };\r\n  },\r\n\r\n  computed: {\r\n    globalViewCompleted: function() {\r\n      return this.completed.header && this.tab_index === this.global_view_index;\r\n    }\r\n  },\r\n\r\n  methods: {\r\n    discardQuotation: function() {\r\n      this.http\r\n        .delete(`/quotations/${this.quotation.id}`)\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.hideDiscardModal();\r\n            this.$router.push({ path: \"/index\" });\r\n            this.alert(this.translations.notifications.quotation_discarded);\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n    },\r\n\r\n    activateQuotation: function() {\r\n      this.http\r\n        .put(`api/quotations/${this.quotation.id}/activate`)\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.alert(\r\n              this.translations.notifications.quotation_activated,\r\n              \"success\"\r\n            );\r\n            this.$router.push({ path: `/${this.quotation.id}/show` });\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        })\r\n        .catch(err => {\r\n          console.log(JSON.stringify(err));\r\n        });\r\n    },\r\n\r\n    hideDiscardModal: function() {\r\n      this.$bvModal.hide(\"discard-modal\");\r\n    }\r\n  },\r\n\r\n  watch: {\r\n    \"completed.header\": function() {\r\n      if (this.completed.header) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.conditions\": function() {\r\n      if (this.completed.conditions) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.services\": function() {\r\n      if (this.completed.services) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.products\": function() {\r\n      if (this.completed.products) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    }\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/edit.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _form_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./form.vue */ \"./viewsrc/modules/quotations/form.vue\");\n\r\n\r\n\r\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\r\n  components: {\r\n    \"quotation-form\": _form_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"]\r\n  },\r\n\r\n  data() {\r\n    return {\r\n      translations: I18n.t(\"quotations.new_edit\")\r\n    };\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/edit.vue?./node_modules/vue-loader/lib??vue-loader-options");
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/form.vue?vue&type=script&lang=js&":
+/*!************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib??vue-loader-options!./viewsrc/modules/quotations/form.vue?vue&type=script&lang=js& ***!
+  \************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _components_header_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./components/header.vue */ \"./viewsrc/modules/quotations/components/header.vue\");\n/* harmony import */ var _components_products_vue__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/products.vue */ \"./viewsrc/modules/quotations/components/products.vue\");\n/* harmony import */ var _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/conditions.vue */ \"./viewsrc/modules/quotations/components/conditions.vue\");\n/* harmony import */ var _components_services_vue__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/services.vue */ \"./viewsrc/modules/quotations/components/services.vue\");\n/* harmony import */ var _components_global_view_vue__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./components/global_view.vue */ \"./viewsrc/modules/quotations/components/global_view.vue\");\n/* harmony import */ var _components_attachments_vue__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./components/attachments.vue */ \"./viewsrc/modules/quotations/components/attachments.vue\");\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\r\n  components: {\r\n    \"quotation-header\": _components_header_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"],\r\n    \"quotation-products\": _components_products_vue__WEBPACK_IMPORTED_MODULE_1__[\"default\"],\r\n    \"quotation-services\": _components_services_vue__WEBPACK_IMPORTED_MODULE_3__[\"default\"],\r\n    \"quotation-conditions\": _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__[\"default\"],\r\n    \"quotation-global-view\": _components_global_view_vue__WEBPACK_IMPORTED_MODULE_4__[\"default\"],\r\n    \"quotation-attachments\": _components_attachments_vue__WEBPACK_IMPORTED_MODULE_5__[\"default\"]\r\n  },\r\n\r\n  props:{\r\n    is_edit: {\r\n      type: Boolean,\r\n      default: false\r\n    }\r\n  },\r\n\r\n  data() {\r\n    return {\r\n      translations: I18n.t(\"quotations.new_edit\"),\r\n      global_view_index: 4,\r\n      tab_index: 0,\r\n      //these are only used to change the color of the tabs to green when the user finishes a section\r\n      completed: {\r\n        header: false,\r\n        products: false,\r\n        services: false,\r\n        conditions: false,\r\n        global_view: false\r\n      },\r\n      refresh_global_view: false,\r\n      quotation: {\r\n        id: parseInt(this.$route.params.id) || null\r\n      }\r\n    };\r\n  },\r\n\r\n  created(){\r\n    if(this.is_edit){\r\n      this.completed.header = true;\r\n      this.completed.products = true;\r\n      this.completed.services = true;\r\n      this.completed.conditions = true;\r\n    }\r\n  },\r\n\r\n  computed: {\r\n    globalViewCompleted: function() {\r\n      return this.completed.header && this.tab_index === this.global_view_index;\r\n    }\r\n  },\r\n\r\n  methods: {\r\n    discardQuotation: function() {\r\n      this.http\r\n        .delete(`/quotations/${this.quotation.id}`)\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.hideDiscardModal();\r\n            this.$router.push({ path: \"/index\" });\r\n            this.alert(this.translations.notifications.quotation_discarded);\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n    },\r\n\r\n    activateQuotation: function() {\r\n      this.http\r\n        .put(`api/quotations/${this.quotation.id}/activate`)\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.alert(\r\n              this.translations.notifications.quotation_activated,\r\n              \"success\"\r\n            );\r\n            this.$router.push({ path: `/${this.quotation.id}/show` });\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        })\r\n        .catch(err => {\r\n          console.log(JSON.stringify(err));\r\n        });\r\n    },\r\n\r\n    hideDiscardModal: function() {\r\n      this.$bvModal.hide(\"discard-modal\");\r\n    }\r\n  },\r\n\r\n  watch: {\r\n    \"completed.header\": function() {\r\n      if (this.completed.header) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.conditions\": function() {\r\n      if (this.completed.conditions) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.services\": function() {\r\n      if (this.completed.services) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.products\": function() {\r\n      if (this.completed.products) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    }\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/form.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3541,7 +3553,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _components_header_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./components/header.vue */ \"./viewsrc/modules/quotations/components/header.vue\");\n/* harmony import */ var _components_products_vue__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/products.vue */ \"./viewsrc/modules/quotations/components/products.vue\");\n/* harmony import */ var _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/conditions.vue */ \"./viewsrc/modules/quotations/components/conditions.vue\");\n/* harmony import */ var _components_services_vue__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/services.vue */ \"./viewsrc/modules/quotations/components/services.vue\");\n/* harmony import */ var _components_global_view_vue__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./components/global_view.vue */ \"./viewsrc/modules/quotations/components/global_view.vue\");\n/* harmony import */ var _components_attachments_vue__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./components/attachments.vue */ \"./viewsrc/modules/quotations/components/attachments.vue\");\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\r\n  components: {\r\n    \"quotation-header\": _components_header_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"],\r\n    \"quotation-products\": _components_products_vue__WEBPACK_IMPORTED_MODULE_1__[\"default\"],\r\n    \"quotation-services\": _components_services_vue__WEBPACK_IMPORTED_MODULE_3__[\"default\"],\r\n    \"quotation-conditions\": _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__[\"default\"],\r\n    \"quotation-global-view\": _components_global_view_vue__WEBPACK_IMPORTED_MODULE_4__[\"default\"],\r\n    \"quotation-attachments\": _components_attachments_vue__WEBPACK_IMPORTED_MODULE_5__[\"default\"]\r\n  },\r\n\r\n  data() {\r\n    return {\r\n      translations: I18n.t(\"quotations.new_edit\"),\r\n      global_view_index: 4,\r\n      tab_index: 0,\r\n      //these are only used to change the color of the tabs to green when the user finishes a section\r\n      completed: {\r\n        header: false,\r\n        products: false,\r\n        services: false,\r\n        conditions: false,\r\n        global_view: false\r\n      },\r\n      refresh_global_view: false,\r\n      quotation: {\r\n        id: null\r\n      }\r\n    };\r\n  },\r\n\r\n  computed: {\r\n    globalViewCompleted: function() {\r\n      return this.completed.header && this.tab_index === this.global_view_index;\r\n    }\r\n  },\r\n\r\n  methods: {\r\n    discardQuotation: function() {\r\n      this.http\r\n        .delete(`/quotations/${this.quotation.id}`)\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.hideDiscardModal();\r\n            this.$router.push({ path: \"/index\" });\r\n            this.alert(this.translations.notifications.quotation_discarded);\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        });\r\n    },\r\n\r\n    activateQuotation: function() {\r\n      this.http\r\n        .put(`api/quotations/${this.quotation.id}/activate`)\r\n        .then(response => {\r\n          if (response.successful) {\r\n            this.alert(\r\n              this.translations.notifications.quotation_activated,\r\n              \"success\"\r\n            );\r\n            this.$router.push({ path: `/${this.quotation.id}/show` });\r\n          } else {\r\n            this.handleError(response.error);\r\n          }\r\n        })\r\n        .catch(err => {\r\n          console.log(\"Error\", err.stack, err.name, err.message);\r\n        })\r\n        .catch(err => {\r\n          console.log(JSON.stringify(err));\r\n        });\r\n    },\r\n\r\n    hideDiscardModal: function() {\r\n      this.$bvModal.hide(\"discard-modal\");\r\n    }\r\n  },\r\n\r\n  watch: {\r\n    \"completed.header\": function() {\r\n      if (this.completed.header) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.conditions\": function() {\r\n      if (this.completed.conditions) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.services\": function() {\r\n      if (this.completed.services) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    },\r\n    \"completed.products\": function() {\r\n      if (this.completed.products) {\r\n        this.refresh_global_view = true;\r\n        this.tab_index++;\r\n      }\r\n    }\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/new.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _form_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./form.vue */ \"./viewsrc/modules/quotations/form.vue\");\n\r\n\r\n\r\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\r\n  components: {\r\n    \"quotation-form\": _form_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"]\r\n  },\r\n\r\n  data() {\r\n    return {\r\n      translations: I18n.t(\"quotations.new_edit\")\r\n    };\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/new.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3673,7 +3685,19 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\"div\", { staticClass: \"row\" }, [\n    _c(\n      \"div\",\n      { staticClass: \"col-lg-12 offset-xl-1 col-xl-10\" },\n      [\n        _c(\n          \"b-modal\",\n          {\n            attrs: {\n              id: \"discard-modal\",\n              title: _vm.translations.discard_modal.title\n            },\n            scopedSlots: _vm._u([\n              {\n                key: \"modal-footer\",\n                fn: function() {\n                  return [\n                    _c(\n                      \"div\",\n                      [\n                        _c(\n                          \"b-button\",\n                          {\n                            attrs: { variant: \"danger\" },\n                            on: { click: _vm.discardQuotation }\n                          },\n                          [\n                            _vm._v(\n                              _vm._s(_vm.translations.discard_modal.yes_text)\n                            )\n                          ]\n                        ),\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-button\",\n                          {\n                            staticClass: \"btn btn-secondary\",\n                            on: { click: _vm.hideDiscardModal }\n                          },\n                          [\n                            _vm._v(\n                              _vm._s(_vm.translations.discard_modal.no_text)\n                            )\n                          ]\n                        )\n                      ],\n                      1\n                    )\n                  ]\n                },\n                proxy: true\n              }\n            ])\n          },\n          [_c(\"p\", [_vm._v(_vm._s(_vm.translations.discard_modal.body))])]\n        ),\n        _vm._v(\" \"),\n        _c(\"h2\", { staticClass: \"text-primary\" }, [\n          _c(\"b\", [_vm._v(_vm._s(_vm.translations.edit_title))])\n        ]),\n        _vm._v(\" \"),\n        _c(\n          \"div\",\n          [\n            _c(\n              \"b-card\",\n              { attrs: { \"no-body\": \"\" } },\n              [\n                _c(\n                  \"b-tabs\",\n                  {\n                    attrs: { card: \"\" },\n                    model: {\n                      value: _vm.tab_index,\n                      callback: function($$v) {\n                        _vm.tab_index = $$v\n                      },\n                      expression: \"tab_index\"\n                    }\n                  },\n                  [\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.header\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(_vm.translations.header.title) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-header\", {\n                              attrs: {\n                                section_valid: _vm.completed.header,\n                                quotation_id: _vm.quotation.id,\n                                get_header: true\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"header\",\n                                    $event\n                                  )\n                                },\n                                \"update:quotation_id\": function($event) {\n                                  return _vm.$set(_vm.quotation, \"id\", $event)\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.products\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.materials.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-products\", {\n                              attrs: {\n                                section_valid: _vm.completed.products,\n                                quotation_id: _vm.quotation.id,\n                                get_products: true\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"products\",\n                                    $event\n                                  )\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.services\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.services.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-services\", {\n                              attrs: {\n                                section_valid: _vm.completed.services,\n                                quotation_id: _vm.quotation.id,\n                                get_services: true\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"services\",\n                                    $event\n                                  )\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.conditions\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.conditions.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-conditions\", {\n                              attrs: {\n                                section_valid: _vm.completed.conditions,\n                                quotation_id: _vm.quotation.id,\n                                get_conditions: true\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"conditions\",\n                                    $event\n                                  )\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        attrs: { active: \"\" },\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.global\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.global_view.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"div\", { staticClass: \"row\" }, [\n                              _c(\n                                \"div\",\n                                { staticClass: \"col-12 text-right mb-2\" },\n                                [\n                                  _c(\n                                    \"b-button\",\n                                    {\n                                      directives: [\n                                        {\n                                          name: \"b-modal\",\n                                          rawName: \"v-b-modal.discard-modal\",\n                                          modifiers: { \"discard-modal\": true }\n                                        }\n                                      ],\n                                      attrs: {\n                                        variant: \"danger\",\n                                        disabled: _vm.quotation.id == null\n                                      }\n                                    },\n                                    [\n                                      _c(\"i\", { staticClass: \"fas fa-trash\" }),\n                                      _vm._v(\n                                        \"\\n                     \" +\n                                          _vm._s(\n                                            _vm.translations.buttons\n                                              .discard_quotation\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]\n                                  ),\n                                  _vm._v(\" \"),\n                                  _c(\n                                    \"b-button\",\n                                    {\n                                      attrs: {\n                                        variant: \"primary\",\n                                        disabled: _vm.quotation.id == null\n                                      },\n                                      on: { click: _vm.activateQuotation }\n                                    },\n                                    [\n                                      _c(\"i\", {\n                                        staticClass: \"fas fa-check-circle\"\n                                      }),\n                                      _vm._v(\n                                        \"\\n                     \" +\n                                          _vm._s(\n                                            _vm.translations.buttons\n                                              .approve_quotation\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]\n                                  )\n                                ],\n                                1\n                              )\n                            ]),\n                            _vm._v(\" \"),\n                            _c(\"quotation-global-view\", {\n                              attrs: {\n                                section_valid: _vm.completed.global_view,\n                                refresh_quotation: _vm.refresh_global_view,\n                                quotation_id: _vm.quotation.id\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"global_view\",\n                                    $event\n                                  )\n                                },\n                                \"update:refresh_quotation\": function($event) {\n                                  _vm.refresh_global_view = $event\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.global\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.attachments_view\n                                            .title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\"quotation-attachments\", {\n                          attrs: { quotation_id: _vm.quotation.id }\n                        })\n                      ],\n                      1\n                    )\n                  ],\n                  1\n                )\n              ],\n              1\n            )\n          ],\n          1\n        )\n      ],\n      1\n    )\n  ])\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/edit.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\"div\", { staticClass: \"row\" }, [\n    _c(\n      \"div\",\n      { staticClass: \"col-lg-12 offset-xl-1 col-xl-10\" },\n      [\n        _c(\"h2\", { staticClass: \"text-primary\" }, [\n          _c(\"b\", [_vm._v(_vm._s(_vm.translations.edit_title))])\n        ]),\n        _vm._v(\" \"),\n        _c(\"quotation-form\", { attrs: { is_edit: true } })\n      ],\n      1\n    )\n  ])\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/edit.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/form.vue?vue&type=template&id=47a20ad6&":
+/*!**********************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./viewsrc/modules/quotations/form.vue?vue&type=template&id=47a20ad6& ***!
+  \**********************************************************************************************************************************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\n    \"div\",\n    [\n      _c(\n        \"b-modal\",\n        {\n          attrs: {\n            id: \"discard-modal\",\n            title: _vm.translations.discard_modal.title\n          },\n          scopedSlots: _vm._u([\n            {\n              key: \"modal-footer\",\n              fn: function() {\n                return [\n                  _c(\n                    \"div\",\n                    [\n                      _c(\n                        \"b-button\",\n                        {\n                          attrs: { variant: \"danger\" },\n                          on: { click: _vm.discardQuotation }\n                        },\n                        [\n                          _vm._v(\n                            _vm._s(_vm.translations.discard_modal.yes_text)\n                          )\n                        ]\n                      ),\n                      _vm._v(\" \"),\n                      _c(\n                        \"b-button\",\n                        {\n                          staticClass: \"btn btn-secondary\",\n                          on: { click: _vm.hideDiscardModal }\n                        },\n                        [_vm._v(_vm._s(_vm.translations.discard_modal.no_text))]\n                      )\n                    ],\n                    1\n                  )\n                ]\n              },\n              proxy: true\n            }\n          ])\n        },\n        [_c(\"p\", [_vm._v(_vm._s(_vm.translations.discard_modal.body))])]\n      ),\n      _vm._v(\" \"),\n      _c(\n        \"b-card\",\n        { attrs: { \"no-body\": \"\" } },\n        [\n          _c(\n            \"b-tabs\",\n            {\n              attrs: { card: \"\" },\n              model: {\n                value: _vm.tab_index,\n                callback: function($$v) {\n                  _vm.tab_index = $$v\n                },\n                expression: \"tab_index\"\n              }\n            },\n            [\n              _c(\n                \"b-tab\",\n                {\n                  attrs: { active: \"\" },\n                  scopedSlots: _vm._u([\n                    {\n                      key: \"title\",\n                      fn: function() {\n                        return [\n                          _c(\n                            \"span\",\n                            { class: { \"text-success\": _vm.completed.header } },\n                            [\n                              _vm._v(\n                                \"\\n            \" +\n                                  _vm._s(_vm.translations.header.title) +\n                                  \"  \\n            \"\n                              ),\n                              _c(\"i\", { staticClass: \"fas fa-check-circle\" })\n                            ]\n                          )\n                        ]\n                      },\n                      proxy: true\n                    }\n                  ])\n                },\n                [\n                  _vm._v(\" \"),\n                  _c(\n                    \"b-card-text\",\n                    [\n                      _c(\"quotation-header\", {\n                        attrs: {\n                          section_valid: _vm.completed.header,\n                          quotation_id: _vm.quotation.id\n                        },\n                        on: {\n                          \"update:section_valid\": function($event) {\n                            return _vm.$set(_vm.completed, \"header\", $event)\n                          },\n                          \"update:quotation_id\": function($event) {\n                            return _vm.$set(_vm.quotation, \"id\", $event)\n                          }\n                        }\n                      })\n                    ],\n                    1\n                  )\n                ],\n                1\n              ),\n              _vm._v(\" \"),\n              _c(\n                \"b-tab\",\n                {\n                  scopedSlots: _vm._u([\n                    {\n                      key: \"title\",\n                      fn: function() {\n                        return [\n                          _c(\n                            \"span\",\n                            {\n                              class: { \"text-success\": _vm.completed.products }\n                            },\n                            [\n                              _vm._v(\n                                \"\\n            \" +\n                                  _vm._s(_vm.translations.materials.title) +\n                                  \"  \\n            \"\n                              ),\n                              _c(\"i\", { staticClass: \"fas fa-check-circle\" })\n                            ]\n                          )\n                        ]\n                      },\n                      proxy: true\n                    }\n                  ])\n                },\n                [\n                  _vm._v(\" \"),\n                  _c(\n                    \"b-card-text\",\n                    [\n                      _c(\"quotation-products\", {\n                        attrs: {\n                          section_valid: _vm.completed.products,\n                          quotation_id: _vm.quotation.id\n                        },\n                        on: {\n                          \"update:section_valid\": function($event) {\n                            return _vm.$set(_vm.completed, \"products\", $event)\n                          }\n                        }\n                      })\n                    ],\n                    1\n                  )\n                ],\n                1\n              ),\n              _vm._v(\" \"),\n              _c(\n                \"b-tab\",\n                {\n                  scopedSlots: _vm._u([\n                    {\n                      key: \"title\",\n                      fn: function() {\n                        return [\n                          _c(\n                            \"span\",\n                            {\n                              class: { \"text-success\": _vm.completed.services }\n                            },\n                            [\n                              _vm._v(\n                                \"\\n            \" +\n                                  _vm._s(_vm.translations.services.title) +\n                                  \"  \\n            \"\n                              ),\n                              _c(\"i\", { staticClass: \"fas fa-check-circle\" })\n                            ]\n                          )\n                        ]\n                      },\n                      proxy: true\n                    }\n                  ])\n                },\n                [\n                  _vm._v(\" \"),\n                  _c(\n                    \"b-card-text\",\n                    [\n                      _c(\"quotation-services\", {\n                        attrs: {\n                          section_valid: _vm.completed.services,\n                          quotation_id: _vm.quotation.id\n                        },\n                        on: {\n                          \"update:section_valid\": function($event) {\n                            return _vm.$set(_vm.completed, \"services\", $event)\n                          }\n                        }\n                      })\n                    ],\n                    1\n                  )\n                ],\n                1\n              ),\n              _vm._v(\" \"),\n              _c(\n                \"b-tab\",\n                {\n                  scopedSlots: _vm._u([\n                    {\n                      key: \"title\",\n                      fn: function() {\n                        return [\n                          _c(\n                            \"span\",\n                            {\n                              class: {\n                                \"text-success\": _vm.completed.conditions\n                              }\n                            },\n                            [\n                              _vm._v(\n                                \"\\n            \" +\n                                  _vm._s(_vm.translations.conditions.title) +\n                                  \"  \\n            \"\n                              ),\n                              _c(\"i\", { staticClass: \"fas fa-check-circle\" })\n                            ]\n                          )\n                        ]\n                      },\n                      proxy: true\n                    }\n                  ])\n                },\n                [\n                  _vm._v(\" \"),\n                  _c(\n                    \"b-card-text\",\n                    [\n                      _c(\"quotation-conditions\", {\n                        attrs: {\n                          section_valid: _vm.completed.conditions,\n                          quotation_id: _vm.quotation.id,\n                          get_conditions: true\n                        },\n                        on: {\n                          \"update:section_valid\": function($event) {\n                            return _vm.$set(_vm.completed, \"conditions\", $event)\n                          }\n                        }\n                      })\n                    ],\n                    1\n                  )\n                ],\n                1\n              ),\n              _vm._v(\" \"),\n              _c(\n                \"b-tab\",\n                {\n                  scopedSlots: _vm._u([\n                    {\n                      key: \"title\",\n                      fn: function() {\n                        return [\n                          _c(\n                            \"span\",\n                            { class: { \"text-success\": _vm.completed.global } },\n                            [\n                              _vm._v(\n                                \"\\n            \" +\n                                  _vm._s(_vm.translations.global_view.title) +\n                                  \"  \\n            \"\n                              ),\n                              _c(\"i\", { staticClass: \"fas fa-check-circle\" })\n                            ]\n                          )\n                        ]\n                      },\n                      proxy: true\n                    }\n                  ])\n                },\n                [\n                  _vm._v(\" \"),\n                  _c(\n                    \"b-card-text\",\n                    [\n                      _c(\"div\", { staticClass: \"row\" }, [\n                        _c(\n                          \"div\",\n                          { staticClass: \"col-12 text-right mb-2\" },\n                          [\n                            _c(\n                              \"b-button\",\n                              {\n                                directives: [\n                                  {\n                                    name: \"b-modal\",\n                                    rawName: \"v-b-modal.discard-modal\",\n                                    modifiers: { \"discard-modal\": true }\n                                  }\n                                ],\n                                attrs: {\n                                  variant: \"danger\",\n                                  disabled: _vm.quotation.id == null\n                                }\n                              },\n                              [\n                                _c(\"i\", { staticClass: \"fas fa-trash\" }),\n                                _vm._v(\n                                  \"\\n                 \" +\n                                    _vm._s(\n                                      _vm.translations.buttons.discard_quotation\n                                    ) +\n                                    \"\\n              \"\n                                )\n                              ]\n                            ),\n                            _vm._v(\" \"),\n                            _c(\n                              \"b-button\",\n                              {\n                                attrs: {\n                                  variant: \"primary\",\n                                  disabled: _vm.quotation.id == null\n                                },\n                                on: { click: _vm.activateQuotation }\n                              },\n                              [\n                                _c(\"i\", { staticClass: \"fas fa-check-circle\" }),\n                                _vm._v(\n                                  \"\\n                 \" +\n                                    _vm._s(\n                                      _vm.translations.buttons.approve_quotation\n                                    ) +\n                                    \"\\n              \"\n                                )\n                              ]\n                            )\n                          ],\n                          1\n                        )\n                      ]),\n                      _vm._v(\" \"),\n                      _c(\"quotation-global-view\", {\n                        attrs: {\n                          section_valid: _vm.completed.global_view,\n                          refresh_quotation: _vm.refresh_global_view,\n                          quotation_id: _vm.quotation.id\n                        },\n                        on: {\n                          \"update:section_valid\": function($event) {\n                            return _vm.$set(\n                              _vm.completed,\n                              \"global_view\",\n                              $event\n                            )\n                          },\n                          \"update:refresh_quotation\": function($event) {\n                            _vm.refresh_global_view = $event\n                          }\n                        }\n                      })\n                    ],\n                    1\n                  )\n                ],\n                1\n              ),\n              _vm._v(\" \"),\n              _c(\n                \"b-tab\",\n                {\n                  scopedSlots: _vm._u([\n                    {\n                      key: \"title\",\n                      fn: function() {\n                        return [\n                          _c(\n                            \"span\",\n                            { class: { \"text-success\": _vm.completed.global } },\n                            [\n                              _vm._v(\n                                \"\\n            \" +\n                                  _vm._s(\n                                    _vm.translations.attachments_view.title\n                                  ) +\n                                  \"  \\n            \"\n                              ),\n                              _c(\"i\", { staticClass: \"fas fa-check-circle\" })\n                            ]\n                          )\n                        ]\n                      },\n                      proxy: true\n                    }\n                  ])\n                },\n                [\n                  _vm._v(\" \"),\n                  _c(\"quotation-attachments\", {\n                    attrs: { quotation_id: _vm.quotation.id }\n                  })\n                ],\n                1\n              )\n            ],\n            1\n          )\n        ],\n        1\n      )\n    ],\n    1\n  )\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/form.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3697,7 +3721,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\"div\", { staticClass: \"row\" }, [\n    _c(\n      \"div\",\n      { staticClass: \"col-lg-12 offset-xl-1 col-xl-10\" },\n      [\n        _c(\n          \"b-modal\",\n          {\n            attrs: {\n              id: \"discard-modal\",\n              title: _vm.translations.discard_modal.title\n            },\n            scopedSlots: _vm._u([\n              {\n                key: \"modal-footer\",\n                fn: function() {\n                  return [\n                    _c(\n                      \"div\",\n                      [\n                        _c(\n                          \"b-button\",\n                          {\n                            attrs: { variant: \"danger\" },\n                            on: { click: _vm.discardQuotation }\n                          },\n                          [\n                            _vm._v(\n                              _vm._s(_vm.translations.discard_modal.yes_text)\n                            )\n                          ]\n                        ),\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-button\",\n                          {\n                            staticClass: \"btn btn-secondary\",\n                            on: { click: _vm.hideDiscardModal }\n                          },\n                          [\n                            _vm._v(\n                              _vm._s(_vm.translations.discard_modal.no_text)\n                            )\n                          ]\n                        )\n                      ],\n                      1\n                    )\n                  ]\n                },\n                proxy: true\n              }\n            ])\n          },\n          [_c(\"p\", [_vm._v(_vm._s(_vm.translations.discard_modal.body))])]\n        ),\n        _vm._v(\" \"),\n        _c(\"h2\", { staticClass: \"text-primary\" }, [\n          _c(\"b\", [_vm._v(_vm._s(_vm.translations.new_title))])\n        ]),\n        _vm._v(\" \"),\n        _c(\n          \"div\",\n          [\n            _c(\n              \"b-card\",\n              { attrs: { \"no-body\": \"\" } },\n              [\n                _c(\n                  \"b-tabs\",\n                  {\n                    attrs: { card: \"\" },\n                    model: {\n                      value: _vm.tab_index,\n                      callback: function($$v) {\n                        _vm.tab_index = $$v\n                      },\n                      expression: \"tab_index\"\n                    }\n                  },\n                  [\n                    _c(\n                      \"b-tab\",\n                      {\n                        attrs: { active: \"\" },\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.header\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(_vm.translations.header.title) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-header\", {\n                              attrs: {\n                                section_valid: _vm.completed.header,\n                                quotation_id: _vm.quotation.id\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"header\",\n                                    $event\n                                  )\n                                },\n                                \"update:quotation_id\": function($event) {\n                                  return _vm.$set(_vm.quotation, \"id\", $event)\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.products\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.materials.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-products\", {\n                              attrs: {\n                                section_valid: _vm.completed.products,\n                                quotation_id: _vm.quotation.id\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"products\",\n                                    $event\n                                  )\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.services\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.services.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-services\", {\n                              attrs: {\n                                section_valid: _vm.completed.services,\n                                quotation_id: _vm.quotation.id\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"services\",\n                                    $event\n                                  )\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.conditions\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.conditions.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"quotation-conditions\", {\n                              attrs: {\n                                section_valid: _vm.completed.conditions,\n                                quotation_id: _vm.quotation.id\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"conditions\",\n                                    $event\n                                  )\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.global\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.global_view.title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\n                          \"b-card-text\",\n                          [\n                            _c(\"div\", { staticClass: \"row\" }, [\n                              _c(\n                                \"div\",\n                                { staticClass: \"col-12 text-right mb-2\" },\n                                [\n                                  _c(\n                                    \"b-button\",\n                                    {\n                                      directives: [\n                                        {\n                                          name: \"b-modal\",\n                                          rawName: \"v-b-modal.discard-modal\",\n                                          modifiers: { \"discard-modal\": true }\n                                        }\n                                      ],\n                                      attrs: {\n                                        variant: \"danger\",\n                                        disabled: _vm.quotation.id == null\n                                      }\n                                    },\n                                    [\n                                      _c(\"i\", { staticClass: \"fas fa-trash\" }),\n                                      _vm._v(\n                                        \"\\n                     \" +\n                                          _vm._s(\n                                            _vm.translations.buttons\n                                              .discard_quotation\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]\n                                  ),\n                                  _vm._v(\" \"),\n                                  _c(\n                                    \"b-button\",\n                                    {\n                                      attrs: {\n                                        variant: \"primary\",\n                                        disabled: _vm.quotation.id == null\n                                      },\n                                      on: { click: _vm.activateQuotation }\n                                    },\n                                    [\n                                      _c(\"i\", {\n                                        staticClass: \"fas fa-check-circle\"\n                                      }),\n                                      _vm._v(\n                                        \"\\n                     \" +\n                                          _vm._s(\n                                            _vm.translations.buttons\n                                              .approve_quotation\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]\n                                  )\n                                ],\n                                1\n                              )\n                            ]),\n                            _vm._v(\" \"),\n                            _c(\"quotation-global-view\", {\n                              attrs: {\n                                section_valid: _vm.completed.global_view,\n                                refresh_quotation: _vm.refresh_global_view,\n                                quotation_id: _vm.quotation.id\n                              },\n                              on: {\n                                \"update:section_valid\": function($event) {\n                                  return _vm.$set(\n                                    _vm.completed,\n                                    \"global_view\",\n                                    $event\n                                  )\n                                },\n                                \"update:refresh_quotation\": function($event) {\n                                  _vm.refresh_global_view = $event\n                                }\n                              }\n                            })\n                          ],\n                          1\n                        )\n                      ],\n                      1\n                    ),\n                    _vm._v(\" \"),\n                    _c(\n                      \"b-tab\",\n                      {\n                        scopedSlots: _vm._u([\n                          {\n                            key: \"title\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"span\",\n                                  {\n                                    class: {\n                                      \"text-success\": _vm.completed.global\n                                    }\n                                  },\n                                  [\n                                    _vm._v(\n                                      \"\\n                \" +\n                                        _vm._s(\n                                          _vm.translations.attachments_view\n                                            .title\n                                        ) +\n                                        \"  \\n                \"\n                                    ),\n                                    _c(\"i\", {\n                                      staticClass: \"fas fa-check-circle\"\n                                    })\n                                  ]\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ])\n                      },\n                      [\n                        _vm._v(\" \"),\n                        _c(\"quotation-attachments\", {\n                          attrs: { quotation_id: _vm.quotation.id }\n                        })\n                      ],\n                      1\n                    )\n                  ],\n                  1\n                )\n              ],\n              1\n            )\n          ],\n          1\n        )\n      ],\n      1\n    )\n  ])\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/new.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\"div\", { staticClass: \"row\" }, [\n    _c(\n      \"div\",\n      { staticClass: \"col-lg-12 offset-xl-1 col-xl-10\" },\n      [\n        _c(\"h2\", { staticClass: \"text-primary\" }, [\n          _c(\"b\", [_vm._v(_vm._s(_vm.translations.new_title))])\n        ]),\n        _vm._v(\" \"),\n        _c(\"quotation-form\")\n      ],\n      1\n    )\n  ])\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/new.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -4140,6 +4164,42 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _nod
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_vue_vue_type_template_id_ea394dc8___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../node_modules/vue-loader/lib??vue-loader-options!./edit.vue?vue&type=template&id=ea394dc8& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/edit.vue?vue&type=template&id=ea394dc8&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_vue_vue_type_template_id_ea394dc8___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_vue_vue_type_template_id_ea394dc8___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/edit.vue?");
+
+/***/ }),
+
+/***/ "./viewsrc/modules/quotations/form.vue":
+/*!*********************************************!*\
+  !*** ./viewsrc/modules/quotations/form.vue ***!
+  \*********************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _form_vue_vue_type_template_id_47a20ad6___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./form.vue?vue&type=template&id=47a20ad6& */ \"./viewsrc/modules/quotations/form.vue?vue&type=template&id=47a20ad6&\");\n/* harmony import */ var _form_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./form.vue?vue&type=script&lang=js& */ \"./viewsrc/modules/quotations/form.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ \"./node_modules/vue-loader/lib/runtime/componentNormalizer.js\");\n\n\n\n\n\n/* normalize component */\n\nvar component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(\n  _form_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n  _form_vue_vue_type_template_id_47a20ad6___WEBPACK_IMPORTED_MODULE_0__[\"render\"],\n  _form_vue_vue_type_template_id_47a20ad6___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"],\n  false,\n  null,\n  null,\n  null\n  \n)\n\n/* hot reload */\nif (false) { var api; }\ncomponent.options.__file = \"viewsrc/modules/quotations/form.vue\"\n/* harmony default export */ __webpack_exports__[\"default\"] = (component.exports);\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/form.vue?");
+
+/***/ }),
+
+/***/ "./viewsrc/modules/quotations/form.vue?vue&type=script&lang=js&":
+/*!**********************************************************************!*\
+  !*** ./viewsrc/modules/quotations/form.vue?vue&type=script&lang=js& ***!
+  \**********************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_index_js_vue_loader_options_form_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib??vue-loader-options!./form.vue?vue&type=script&lang=js& */ \"./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/form.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__[\"default\"] = (_node_modules_vue_loader_lib_index_js_vue_loader_options_form_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[\"default\"]); \n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/form.vue?");
+
+/***/ }),
+
+/***/ "./viewsrc/modules/quotations/form.vue?vue&type=template&id=47a20ad6&":
+/*!****************************************************************************!*\
+  !*** ./viewsrc/modules/quotations/form.vue?vue&type=template&id=47a20ad6& ***!
+  \****************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_form_vue_vue_type_template_id_47a20ad6___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../node_modules/vue-loader/lib??vue-loader-options!./form.vue?vue&type=template&id=47a20ad6& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/form.vue?vue&type=template&id=47a20ad6&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_form_vue_vue_type_template_id_47a20ad6___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_form_vue_vue_type_template_id_47a20ad6___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/form.vue?");
 
 /***/ }),
